requirements.txt
einops==0.8.0
kornia==0.8.0
tqdm==4.67.1
matplotlib==3.10.0
opencv-python==4.11.0.86
transformers==4.41.2
scipy==1.15.1
moderngl==5.12.0
pyopengl==3.1.9
glfw==2.8.0
imgui[glfw]==2.0.0
geopandas==1.0.1
pyntcloud==0.3.1
roma==1.5.1
neptune==1.13.0
python-dotenv==1.0.1
pykdtree==1.4.1
torchmetrics==1.6.1
openpyxl==3.1.5
laspy[lazrs]==2.5.4
e3nn==0.5.5
edt_gaussian_rasterization @ git+https://github.com/geospaitial-lab/EasyDigiTwin-gaussian-rasterization.git
torch2moderngl @ git+https://github.com/geospaitial-lab/torch2moderngl.git